#10352: Update namespaces, locations, and code order of matmul
