+++ tests/ttnn/unit_tests/gtests/test_to_and_from_json.cpp
@@ -27,30 +27,30 @@ TEST(TEST_JSON_CONVERSION, TEST_MEMORY_CONFIG) {
 
 TEST(TEST_JSON_CONVERSION, TEST_MATMUL_CONFIG) {
     auto matmul_multi_core_reuse_program_config =
-        ttnn::MatmulMultiCoreReuseProgramConfig{CoreCoord{2, 3}, 32, 64, 48, 128, 96};
-    auto matmul_program_config = ttnn::MatmulProgramConfig{matmul_multi_core_reuse_program_config};
+        ttnn::operations::matmul::MatmulMultiCoreReuseProgramConfig{CoreCoord{2, 3}, 32, 64, 48, 128, 96};
+    auto matmul_program_config = ttnn::operations::matmul::MatmulProgramConfig{matmul_multi_core_reuse_program_config};
 
     auto json_object = tt::stl::json::to_json(matmul_program_config);
 
-    auto deserialized_matmul_program_config = tt::stl::json::from_json<ttnn::MatmulProgramConfig>(json_object);
+    auto deserialized_matmul_program_config = tt::stl::json::from_json<ttnn::operations::matmul::MatmulProgramConfig>(json_object);
 
     ASSERT_EQ(
         matmul_multi_core_reuse_program_config.compute_with_storage_grid_size,
-        std::get<ttnn::MatmulMultiCoreReuseProgramConfig>(deserialized_matmul_program_config)
+        std::get<ttnn::operations::matmul::MatmulMultiCoreReuseProgramConfig>(deserialized_matmul_program_config)
             .compute_with_storage_grid_size);
     ASSERT_EQ(
         matmul_multi_core_reuse_program_config.in0_block_w,
-        std::get<ttnn::MatmulMultiCoreReuseProgramConfig>(deserialized_matmul_program_config).in0_block_w);
+        std::get<ttnn::operations::matmul::MatmulMultiCoreReuseProgramConfig>(deserialized_matmul_program_config).in0_block_w);
     ASSERT_EQ(
         matmul_multi_core_reuse_program_config.out_subblock_h,
-        std::get<ttnn::MatmulMultiCoreReuseProgramConfig>(deserialized_matmul_program_config).out_subblock_h);
+        std::get<ttnn::operations::matmul::MatmulMultiCoreReuseProgramConfig>(deserialized_matmul_program_config).out_subblock_h);
     ASSERT_EQ(
         matmul_multi_core_reuse_program_config.out_subblock_w,
-        std::get<ttnn::MatmulMultiCoreReuseProgramConfig>(deserialized_matmul_program_config).out_subblock_w);
+        std::get<ttnn::operations::matmul::MatmulMultiCoreReuseProgramConfig>(deserialized_matmul_program_config).out_subblock_w);
     ASSERT_EQ(
         matmul_multi_core_reuse_program_config.per_core_M,
-        std::get<ttnn::MatmulMultiCoreReuseProgramConfig>(deserialized_matmul_program_config).per_core_M);
+        std::get<ttnn::operations::matmul::MatmulMultiCoreReuseProgramConfig>(deserialized_matmul_program_config).per_core_M);
     ASSERT_EQ(
         matmul_multi_core_reuse_program_config.per_core_N,
-        std::get<ttnn::MatmulMultiCoreReuseProgramConfig>(deserialized_matmul_program_config).per_core_N);
+        std::get<ttnn::operations::matmul::MatmulMultiCoreReuseProgramConfig>(deserialized_matmul_program_config).per_core_N);
 }
